refactor: lift resolveEffectiveConfig out of root main.go into configresolve/
Moves the ~940-line precedence-aware config merger (CLI > env > YAML >
defaults) and its CLIInputs / Resolved types out of the all-in-one
duckgres binary's package into a new configresolve/ package. This is the
last piece needed before cmd/duckgres-controlplane and cmd/duckgres-worker
can move past their config-load stub and actually boot — they can now
import and call configresolve.ResolveEffective(...) instead of having to
duplicate the resolution logic.

Why a new package and not configloader/: configloader is intentionally
light (YAML schema + LoadFile only, ~30 lines per file, no deps on
server/controlplane). configresolve is heavy by design (depends on
server, controlplane, server/ducklake) because it produces fully wired
runtime configs. Mixing the two would force everything that just wants
the YAML schema to pull in the heavy graph too.

Naming changes (all renames mechanical, no logic touched):
- configCLIInputs → configresolve.CLIInputs
- resolvedConfig → configresolve.Resolved
- defaultServerConfig → configresolve.DefaultServerConfig
- resolveEffectiveConfig → configresolve.ResolveEffective
- splitAndTrim, intPtr, boolPtr stay package-private (unexported)

main.go's existing FileConfig type alias to configloader.FileConfig is
unchanged; ResolveEffective takes *configloader.FileConfig directly so
callers don't need the alias.

Verified:
- go build ./... + go build -tags kubernetes ./... both clean
- go test ./configresolve/... passes both with and without -tags kubernetes
- go list -deps ./cmd/duckgres-controlplane confirms libduckdb still NOT
  pulled in (default tags AND -tags kubernetes)

The follow-up wiring of cmd/duckgres-controlplane and cmd/duckgres-worker
to actually call configresolve.ResolveEffective lands in a separate PR
since each binary needs its own flag.* declarations + cliSet capture
shape; doing both here would balloon this PR.
